refactor: prep admin rework deductions extraction
